implement LoggerClient library
