+++ src/CMakeLists.txt
@@ -12,3 +12,4 @@ add_subdirectory(libWhole_body_controllers)
 add_subdirectory(libTrajectory_planner)
 add_subdirectory(libKinDyn_wrapper)
 add_subdirectory(libRetargeting_helper)
+add_subdirectory(libLogger_client)
